// pro/src/apiClient/v1/models/SubcategoryIdEnum.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

/**
 * An enumeration.
 */
export enum SubcategoryIdEnum {
  ABO_BIBLIOTHEQUE = 'ABO_BIBLIOTHEQUE',
  ABO_CONCERT = 'ABO_CONCERT',
  ABO_JEU_VIDEO = 'ABO_JEU_VIDEO',
  ABO_LIVRE_NUMERIQUE = 'ABO_LIVRE_NUMERIQUE',
  ABO_LUDOTHEQUE = 'ABO_LUDOTHEQUE',
  ABO_MEDIATHEQUE = 'ABO_MEDIATHEQUE',
  ABO_MUSEE = 'ABO_MUSEE',
  ABO_PLATEFORME_MUSIQUE = 'ABO_PLATEFORME_MUSIQUE',
  ABO_PLATEFORME_VIDEO = 'ABO_PLATEFORME_VIDEO',
  ABO_PRATIQUE_ART = 'ABO_PRATIQUE_ART',
  ABO_PRESSE_EN_LIGNE = 'ABO_PRESSE_EN_LIGNE',
  ABO_SPECTACLE = 'ABO_SPECTACLE',
  ACHAT_INSTRUMENT = 'ACHAT_INSTRUMENT',
  ACTIVATION_EVENT = 'ACTIVATION_EVENT',
  ACTIVATION_THING = 'ACTIVATION_THING',
  APP_CULTURELLE = 'APP_CULTURELLE',
  ATELIER_PRATIQUE_ART = 'ATELIER_PRATIQUE_ART',
  AUTRE_SUPPORT_NUMERIQUE = 'AUTRE_SUPPORT_NUMERIQUE',
  BON_ACHAT_INSTRUMENT = 'BON_ACHAT_INSTRUMENT',
  CAPTATION_MUSIQUE = 'CAPTATION_MUSIQUE',
  CARTE_CINE_ILLIMITE = 'CARTE_CINE_ILLIMITE',
  CARTE_CINE_MULTISEANCES = 'CARTE_CINE_MULTISEANCES',
  CARTE_JEUNES = 'CARTE_JEUNES',
  CARTE_MUSEE = 'CARTE_MUSEE',
  CINE_PLEIN_AIR = 'CINE_PLEIN_AIR',
  CINE_VENTE_DISTANCE = 'CINE_VENTE_DISTANCE',
  CONCERT = 'CONCERT',
  CONCOURS = 'CONCOURS',
  CONFERENCE = 'CONFERENCE',
  DECOUVERTE_METIERS = 'DECOUVERTE_METIERS',
  ESCAPE_GAME = 'ESCAPE_GAME',
  EVENEMENT_CINE = 'EVENEMENT_CINE',
  EVENEMENT_JEU = 'EVENEMENT_JEU',
  EVENEMENT_MUSIQUE = 'EVENEMENT_MUSIQUE',
  EVENEMENT_PATRIMOINE = 'EVENEMENT_PATRIMOINE',
  FESTIVAL_ART_VISUEL = 'FESTIVAL_ART_VISUEL',
  FESTIVAL_CINE = 'FESTIVAL_CINE',
  FESTIVAL_LIVRE = 'FESTIVAL_LIVRE',
  FESTIVAL_MUSIQUE = 'FESTIVAL_MUSIQUE',
  FESTIVAL_SPECTACLE = 'FESTIVAL_SPECTACLE',
  JEU_EN_LIGNE = 'JEU_EN_LIGNE',
  JEU_SUPPORT_PHYSIQUE = 'JEU_SUPPORT_PHYSIQUE',
  LIVESTREAM_EVENEMENT = 'LIVESTREAM_EVENEMENT',
  LIVESTREAM_MUSIQUE = 'LIVESTREAM_MUSIQUE',
  LIVESTREAM_PRATIQUE_ARTISTIQUE = 'LIVESTREAM_PRATIQUE_ARTISTIQUE',
  LIVRE_AUDIO_PHYSIQUE = 'LIVRE_AUDIO_PHYSIQUE',
  LIVRE_NUMERIQUE = 'LIVRE_NUMERIQUE',
  LIVRE_PAPIER = 'LIVRE_PAPIER',
  LOCATION_INSTRUMENT = 'LOCATION_INSTRUMENT',
  MATERIEL_ART_CREATIF = 'MATERIEL_ART_CREATIF',
  MUSEE_VENTE_DISTANCE = 'MUSEE_VENTE_DISTANCE',
  OEUVRE_ART = 'OEUVRE_ART',
  PARTITION = 'PARTITION',
  PLATEFORME_PRATIQUE_ARTISTIQUE = 'PLATEFORME_PRATIQUE_ARTISTIQUE',
  PRATIQUE_ART_VENTE_DISTANCE = 'PRATIQUE_ART_VENTE_DISTANCE',
  PODCAST = 'PODCAST',
  RENCONTRE_EN_LIGNE = 'RENCONTRE_EN_LIGNE',
  RENCONTRE_JEU = 'RENCONTRE_JEU',
  RENCONTRE = 'RENCONTRE',
  SALON = 'SALON',
  SEANCE_CINE = 'SEANCE_CINE',
  SEANCE_ESSAI_PRATIQUE_ART = 'SEANCE_ESSAI_PRATIQUE_ART',
  SPECTACLE_ENREGISTRE = 'SPECTACLE_ENREGISTRE',
  SPECTACLE_REPRESENTATION = 'SPECTACLE_REPRESENTATION',
  SPECTACLE_VENTE_DISTANCE = 'SPECTACLE_VENTE_DISTANCE',
  SUPPORT_PHYSIQUE_FILM = 'SUPPORT_PHYSIQUE_FILM',
  SUPPORT_PHYSIQUE_MUSIQUE = 'SUPPORT_PHYSIQUE_MUSIQUE',
  TELECHARGEMENT_LIVRE_AUDIO = 'TELECHARGEMENT_LIVRE_AUDIO',
  TELECHARGEMENT_MUSIQUE = 'TELECHARGEMENT_MUSIQUE',
  VISITE_GUIDEE = 'VISITE_GUIDEE',
  VISITE_VIRTUELLE = 'VISITE_VIRTUELLE',
  VISITE = 'VISITE',
  VOD = 'VOD',
}